Fixing an error with tinymce changed their naming convention
When editing a user contribution, error occurs when a contributor logs in and uses the edit link to edit a contribution. Issue is due to an update that occurred when wysiwyg editor tiny mce changed their naming convention from tiny_mce to tinymce.

This was resolved with Omeka 2.6 when editors updated plugins, but was NOT updated on the  edit.php file